gateway/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pyserial==3.5
paho-mqtt==2.1.0
pyjwt==2.10.1
pyyaml==6.0.2
